Create generic action element and use on buttons
