+++ services/manage/app/templates/components/action-element.hbs
@@ -1,0 +1,1 @@
+{{yield}}
